+++ tests/test_mask.py
@@ -425,3 +425,27 @@ def test_array_tree_mask_tree_unmask():
     assert not (frozen_array == freeze(arraylib.ones((5, 6))))
     # assert not (frozen_array == freeze(arraylib.ones((5, 5)).astype(arraylib.uint8)))
     assert hash(frozen_array) == hash(frozen_array)
+
+
+def test_custom_mask_unmask_wrappers():
+
+    @dc.dataclass
+    class MyInt:
+        value: int
+    @dc.dataclass
+    class MaskedInt:
+        value: MyInt
+    # define a rule of how to mask instances of MyInt
+    @tree_mask.def_type(MyInt)
+    def mask_int(value):
+        return MaskedInt(value)
+    # define a rule how to unmask the MaskedInt wrapper
+    @tree_unmask.def_type(MaskedInt)
+    def unmask_int(value):
+        return value.value
+    tree = [MyInt(1), MyInt(2), {"a": MyInt(3)}]
+    masked_tree = tree_mask(tree, cond=lambda _: True)
+    masked_tree
+    assert tree_unmask(masked_tree) == [MyInt(value=1), MyInt(value=2), {'a': MyInt(value=3)}]
+    # check the mask type is recognized by is_masked
+    assert is_masked(masked_tree[0]) is True